+++ matchms/similarity/__init__.py
@@ -1,11 +1,13 @@
 """similarity module"""
 from .CosineGreedy import CosineGreedy
+from .CosineHungarian import CosineHungarian
 from .IntersectMz import IntersectMz
 from .ModifiedCosine import ModifiedCosine
 
 
 __all__ = [
     "CosineGreedy",
+    "CosineHungarian",
     "IntersectMz",
     "ModifiedCosine",
 ]