+++ matchms/similarity/__init__.py
@@ -1,5 +1,6 @@
 """similarity module"""
 from .CosineGreedy import CosineGreedy
+from .CosineGreedyNumba import CosineGreedyNumba
 from .CosineHungarian import CosineHungarian
 from .IntersectMz import IntersectMz
 from .ModifiedCosine import ModifiedCosine
@@ -7,6 +8,7 @@ from .ModifiedCosine import ModifiedCosine
 
 __all__ = [
     "CosineGreedy",
+    "CosineGreedyNumba",
     "CosineHungarian",
     "IntersectMz",
     "ModifiedCosine",